+++ modules/core/src/test/kotlin/io/github/lmliam/kotventure/core/selector/SelectorDslTest.kt
@@ -1,5 +1,6 @@
 package io.github.lmliam.kotventure.core.selector
 
+import io.github.lmliam.kotventure.core.nbt.list
 import io.github.lmliam.kotventure.test.text.childAt
 import io.github.lmliam.kotventure.test.text.shouldBeSelectorComponent
 import io.github.lmliam.kotventure.test.text.shouldContainText
@@ -85,6 +86,22 @@ class SelectorDslTest :
                     ).shouldBeSelectorComponent()
 
                 component shouldHaveSelectorPattern "@e[team=!blue,team=!]"
+            }
+
+            "builds a selector component with typed NBT filters" {
+                val component =
+                    selector(
+                        entities {
+                            nbt {
+                                "Tags" eq list("boss")
+                            }
+                            not {
+                                nbt { "Silent" eq true }
+                            }
+                        },
+                    ).shouldBeSelectorComponent()
+
+                component shouldHaveSelectorPattern "@e[nbt={Tags:[\"boss\"]},nbt=!{Silent:1b}]"
             }
 
             "builds a selector component with the escape hatch" {